Move backend objects to dedicated module
