+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+p3
+==
+
+.. toctree::
+   :maxdepth: 4
+
+   p3
