+++ docs/source/p3.data.rst
@@ -5,6 +5,6 @@ Module contents
 ---------------
 
 .. automodule:: p3.data
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
