+++ docs/source/p3.report.rst
@@ -5,6 +5,6 @@ Module contents
 ---------------
 
 .. automodule:: p3.report
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
