+++ docs/source/p3.rst
@@ -5,16 +5,17 @@ Subpackages
 -----------
 
 .. toctree::
+   :maxdepth: 4
 
-    p3.data
-    p3.metrics
-    p3.plot
-    p3.report
+   p3.data
+   p3.metrics
+   p3.plot
+   p3.report
 
 Module contents
 ---------------
 
 .. automodule:: p3
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
