+++ p3/plot/__init__.py
@@ -1,16 +1,10 @@
 # Copyright (c) 2022-2023 Intel Corporation
 # SPDX-License-Identifier: MIT
 
-from p3.plot._cascade import cascade, CascadePlot, CascadePlotMatplotlib
-from p3.plot._navchart import navchart, NavChart, NavChartMatplotlib
-from p3.plot._common import Plot
+from p3.plot._cascade import cascade
+from p3.plot._navchart import navchart
 
 __all__ = [
     "cascade",
     "navchart",
-    "Plot",
-    "CascadePlot",
-    "CascadePlotMatplotlib",
-    "NavChart",
-    "NavChartMatplotlib",
 ]
